Add NewTabPageConfigurationClient
